refactor: seperate the baseball class
